+++ src/test/kotlin/dev/nybroe/collector/inspections/CollectFunctionOnCollectionInspectionTest.kt
@@ -1,0 +1,16 @@
+package dev.nybroe.collector.inspections
+
+import com.intellij.codeInspection.InspectionProfileEntry
+
+internal class CollectFunctionOnCollectionInspectionTest : InspectionTest() {
+    override fun defaultInspection(): InspectionProfileEntry = CollectFunctionInCollectionInspection()
+    override fun defaultAction(): String = ""
+
+    fun testCollectOnCollectVariable() {
+        doTest("collect_on_collect_variable")
+    }
+
+    fun testCollectInCollect() {
+        doTest("collect_in_collect")
+    }
+}