+++ src/test/resources/inspections/CollectFunctionInCollectionInspection/collect_in_collect.php
@@ -1,0 +1,3 @@
+<?php
+
+$sum =  <warning descr="Recursive call on collection">collect(collect([1,2,3]))</warning>->sum();