+++ src/test/resources/inspections/CollectFunctionInCollectionInspection/collect_on_collect_variable.php
@@ -1,0 +1,5 @@
+<?php
+
+$variable = collect([1,2,3]);
+
+$sum = <warning descr="Recursive call on collection">collect($variable)</warning>->sum();